use sha-1 when creating object identities (rather than sha-256) cryptographic secrity is not a requirement, so the extra computational overhead can be avoided.
