+++ agent/lib/router_stats.js
@@ -20,8 +20,7 @@ var rhea = require('rhea');
 var path = require('path');
 var fs = require('fs');
 var Router = require('./qdr.js').Router;
-var crypto = require('crypto');
-var uuidv5 = require('uuid/v5');
+var myutils = require('./utils.js');
 
 
 
@@ -243,20 +242,6 @@ function is_application_connection (c) {
 }
 
 
-function generateStableUuid() {
-    var hash = crypto.createHash('sha256');
-    for (var i = 0, j = arguments.length; i < j; i++){
-        var argument = arguments[i];
-        if (argument) {
-            hash.update(argument);
-        }
-    }
-    var ba = [];
-    ba.push(...hash.digest().slice(0, 16));
-    const ns = "3751f842-240e-48b9-89b5-5b47f04e931b";
-    return uuidv5(ns, ba);
-}
-
 function get_normal_connections (results) {
     var connections = {};
     results.forEach(function (stats, i) {
@@ -268,7 +253,7 @@ function get_normal_connections (results) {
             var addressSpace = process.env.ADDRESS_SPACE;
             var addressSpaceNamespace = process.env.ADDRESS_SPACE_NAMESPACE;
             var addressSpaceType = process.env.ADDRESS_SPACE_TYPE;
-            var uuid = generateStableUuid(addressSpaceNamespace, addressSpace, c.container, c.host);
+            var uuid = myutils.generate_stable_uuid(addressSpaceNamespace, addressSpace, c.container, c.host);
             connections[qualified_id] = {
                 id: c.identity,
                 addressSpace: addressSpace,
